requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
pillow==10.1.0
accelerate==0.25.0
